requirements.txt
pillow
ollama
py3exiv2

moondream

#transformers
#torch
#einops
#pyvips
#pyvips-binary
#torchao